Import OSM Files to Neo4j
Created a Python script to import OSM data to a Neo4j database. Can be used to input large osm.pbf files, but takes a long time to do so.

The Massachusetts dataset takes about 10-20 minutes to complete.

If the import fails, then restart the Neo4j server and retry. If you want to save time, comment out the parts of the code completed successfully in the main function.